refactor: change name of packages in the project
