+++ common/src/test/scala/io/github/positionpal/reaction/ReactionTests.scala
@@ -1,11 +1,11 @@
-package reaction
+package io.github.positionpal.reaction
 
 import scala.language.postfixOps
 
 import cats.{Id, Monad}
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpecLike
-import reaction.Reaction.{Context, OptionReaction}
+import io.github.positionpal.reaction.Reaction.{Context, OptionReaction}
 
 class ReactionTests extends AnyWordSpecLike with Matchers:
