+++ domain/src/main/scala/io/github/positionpal/group/Group.scala
