+++ domain/src/test/scala/io/github/positionpal/group/GroupSpec.scala
@@ -1,9 +1,10 @@
-package messages
+package io.github.positionpal.group
 
 import akka.actor.testkit.typed.scaladsl.ScalaTestWithActorTestKit
 import akka.pattern.StatusReply
 import akka.pattern.StatusReply.Success
 import akka.persistence.testkit.scaladsl.EventSourcedBehaviorTestKit
+import messages.{Command, Event, Group, Reply, UserEnterToGroup, UserLeaveFromGroup}
 import messages.Group.State
 import org.scalatest.BeforeAndAfterEach
 import org.scalatest.wordspec.AnyWordSpecLike
